setting up passive datums

+++ code/modules/antagonists/mind_flayer/flayer_power.dm
@@ -44,7 +44,17 @@
 	flayer.add_ability(path)
 	return TRUE
 
+/datum/mindflayer_passive
+	var/purchase_text = "Oopsie daisies! No purchase text on this ability!"
+	var/upgrade_text = "Uh oh someone forgot to add upgrade text!"
+	///All passives start at level on
+	var/level = 1
+	var/mob/living/owner
+
 /datum/mindflayer_passive/proc/on_apply(datum/antagonist/mindflayer/flayer)
+	return
+
+/datum/mindflayer_passive/proc/on_remove(datum/antagonist/mindflayer/flayer)
 	return
 
 // Retractable weapons code
@@ -76,9 +86,9 @@
 	SIGNAL_HANDLER
 	if(!any_hand && !istype(owner.get_active_hand(), weapon_type))
 		return
-	if(istype(owner.l_hand, weapon_type))
+	if(owner.l_hand && istype(owner.l_hand, weapon_type))
 		qdel(owner.l_hand)
 		owner.update_inv_l_hand()
-	if(istype(owner.r_hand, weapon_type))
+	if(owner.r_hand && istype(owner.r_hand, weapon_type))
 		qdel(owner.r_hand)
 		owner.update_inv_r_hand()
